+++ modules/kubernetes/prometheus/variables.tf
@@ -135,7 +135,7 @@ variable "azad_kube_proxy_enabled" {
 }
 
 # This should be removed, https://github.com/kubernetes/kube-state-metrics/issues/1631
-variable "kube_state_metrics_namepsaces" {
+variable "kube_state_metrics_namespaces" {
   description = "Comma-separated list of namespaces to be enabled. To get metrics from all namespaces use ''"
   type        = string
 }