+++ validation/kubernetes/aks-core/main.tf
@@ -96,5 +96,5 @@ module "aks_core" {
     resource_selector  = ["platform"]
     namespace_selector = ["platform"]
   }
-  kube_state_metrics_namepsaces_extras = ["foo"]
+  kube_state_metrics_namespaces_extras = ["foo"]
 }