+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - [#465](https://github.com/XenitAB/terraform-modules/pull/465) [Breaking] Move xenit credentials to Prometheus and remove Xenit proxy.
 
+### Fixed
+
+- [#466](https://github.com/XenitAB/terraform-modules/pull/466) Replace misspelled variable kube_state_metrics_namepsaces with kube_state_metrics_namespaces
+
 ## 2021.11.6
 
 ### Added